+++ CHANGELOG.md
@@ -1,10 +1,13 @@
 # Changelog
 
-## UNRELEASED
+*Feb 3, 2025*
+
+## v4.2.11
 
 ### Bug Fixes
 
 - [#1099](https://github.com/crypto-org-chain/chain-main/pull/1099) Avoid negative coin amount error when query supply liquid of non BaseCoinUnit.
+- [#]() Update cometbft to v0.34.x.
 
 *Dec 18, 2024*